Move NvtPreferences to web/pages/nvts
Closes #30
