+++ gsa/src/web/pages/scanconfigs/editnvtdetailsdialog.js
@@ -27,30 +27,30 @@ import _ from 'gmp/locale';
 
 import {isDefined} from 'gmp/utils/identity';
 
-import PropTypes from '../../utils/proptypes.js';
+import PropTypes from 'web/utils/proptypes';
 
-import SeverityBar from '../../components/bar/severitybar.js';
+import SeverityBar from 'web/components/bar/severitybar';
 
-import SaveDialog from '../../components/dialog/savedialog.js';
+import SaveDialog from 'web/components/dialog/savedialog';
 
-import Radio from '../../components/form/radio.js';
-import Text from '../../components/form/text.js';
-import TextField from '../../components/form/textfield.js';
+import Radio from 'web/components/form/radio';
+import Text from 'web/components/form/text';
+import TextField from 'web/components/form/textfield';
 
-import Layout from '../../components/layout/layout.js';
+import Layout from 'web/components/layout/layout';
 
-import DetailsLink from '../../components/link/detailslink.js';
-import Link from '../../components/link/link.js';
+import DetailsLink from 'web/components/link/detailslink';
+import Link from 'web/components/link/link';
 
-import SimpleTable from '../../components/table/simpletable.js';
-import Table from '../../components/table/table.js';
-import TableBody from '../../components/table/body.js';
-import TableData from '../../components/table/data.js';
-import TableHead from '../../components/table/head.js';
-import TableHeader from '../../components/table/header.js';
-import TableRow from '../../components/table/row.js';
+import SimpleTable from 'web/components/table/simpletable';
+import Table from 'web/components/table/table';
+import TableBody from 'web/components/table/body';
+import TableData from 'web/components/table/data';
+import TableHead from 'web/components/table/head';
+import TableHeader from 'web/components/table/header';
+import TableRow from 'web/components/table/row';
 
-import NvtPreference from './nvtpreference.js';
+import NvtPreference from '../nvts/nvtpreference';
 
 class EditDialog extends React.Component {
